+++ build.sbt
@@ -1,14 +1,14 @@
 lazy val `dmf-kit` =
   (project in file("."))
     .aggregate(
-      `models-canonical`,
-      `models-common`)
-
-lazy val `models-canonical` =
-  (project in file("modules") / "models" / "canonical")
-    .dependsOn(`models-common`)
+      `models-common`,
+      `models-encoding-border`)
 
 lazy val `models-common` = project in file("modules") / "models" / "common"
+
+lazy val `models-encoding-border` =
+  (project in file("modules") / "models" / "encoding-borer")
+    .dependsOn(`models-common`)
 
 publish / skip := true
