+++ modules/models/encoding-borer/src/main/scala/gig/dmf/canonical/encodings/Canonical.scala
