+++ modules/models/encoding-borer/src/main/scala/gig/dmf/models/encodings/borer/TypedEncoder.scala
@@ -1,4 +1,4 @@
-package gig.dmf.canonical.encodings.cbor
+package gig.dmf.models.encodings.borer
 
 import io.bullet.borer.Encoder
